pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scikg_extract"
version = "1.0.0"
description = "An agentic framework for extracting structured scientific knowledge from scientific literature using LLMs and process schema representation."
authors = [
    {name = "Sameer Sadruddin", email = "sameer.sadruddin@tib.eu"},
    {name = "Jennifer D'Souza", email = "jennifer.dsouza@tib.eu"}
]
maintainers = [
    {name = "Sameer Sadruddin", email = "sameer.sadruddin@tib.eu"}
]
dependencies = [
    "langgraph>=1.0.1",
    "langchain-core>=0.3.74",
    "langchain-openai>=0.3.30",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.2",
    "pandas>=2.2.3",
    "openpyxl>=3.1.5",
    "jsonschema>=4.23.0",
    "PyYAML>=6.0.2",
    "lmdb>=1.7.3",
    "RapidFuzz>=3.14.1",
    "transformers>=4.55.2",
    "docling>=2.28.4",
    "docling-core>=2.25.0",
    "evaluate>=0.4.3"
]
requires-python = ">=3.12"
readme = "README.md"
license = "Apache-2.0"
keywords = ["Structured Scientific Knowledge Extraction", "Scientific Schemas", "Large Language Models", "Agentic Framework"]

[project.urls]
homepage = "https://github.com/sciknoworg/SciKG-extract"
repository = "https://github.com/sciknoworg/SciKG-extract"

[tool.setuptools.packages.find]
include = ["scikg_extract*"]
